[DRAFT] An fts::test that hits indexes
Implement an inlined version of fts::test that succefully hits
indexes. (I tested using imdbench extended with some fts::textsearch
indexes, and added some tests to the explain tests)

Per some discussions with @vpetrovykh, I introduced an fts::make_doc
function that is a variadic optional function that produces a string
that can be reasonable passed to an fts::textsearch index and that
works well will fts::test. (fts::test does the same operations.)

make_doc essentially does `array_to_string("docs", ' ')`, but inlined.

The most fiddly bit about all of this, though, in handling the
language argument to postgres's `to_tsvector`, and how it interacts
with our constant extraction preprocessing. Because we extract string
literals during preprocessing, the `language` argument to `fts::test`
and similar will typically be a parameter.

The language argument is of type `regconfig` (which is an int id);
there is a cast from `text` to `regconfig`, but if the argument to
language is a text query parameter cast to `regconfig`, the index will
not match and the query will be slow.

The index *will* match if the argument is of type `regconfig`, and so
we try to make sure this happens. We introduce a new scalar type,
`fts::language`, that is compiled to `regconfig`.

The binary serialization for `regconfig` is as an int32, so we
introduce machinery to use the text serialization for regconfig params
instead. (Thanks to @elprans for figuring this trick out.)

We then automatically "promote" `str` params that are casted (possibly
implicitly) to `fts::language` to be `fts::language` params, which lets
using a constant literal param to `language` work right.

The bulk of the code here ends up being code to handle introducing our
first type that we want to use text serialization for in the binary
protocol.