Add a variety of cast with a more restricted implicit cast behavior
Currently, implicit casts are applied not just when searching for
functions to apply, but when searching for *casts* to apply.
This is going to cause problems if we add an `fts::language` type with
an implicit cast from str, but as I tried to fix this I discovered
it already *did* cause a problem: it caused us to support casting
from `cal::local_date` to `cal::local_time`, but actually doing so
would produce an ISE.

Add a new category of implicit cast that doesn't have this behavior.

I had originally wanted to get rid of this behavior entirely, which
required adding in a large number of numeric casts, but this then
caused a bunch of knock-on trouble as things that relied on implicit
cast *distance* no longer worked right, which required adding a lot
more overloads. I pushed that WIP to `no-chained-implicit-casts`; it
still needs a dozen or so more overloads of `=` and friends to be made
fully operational.

I could finish with that approach, if that seems better to people, or
maybe someone has a better idea.