Make CompilerContextLevel.new use self as the previous level
Currently in `new` and friends on context levels, self is passed
through as `prevlevel` to `CompilerContext.push`, where it is ignored
and the current most recent level is used instead.

It seems like this behavior may have been introduced by accident in
2019 while adding type annotations 854a52d673, but of course much code
grew to depend on it.

I lost a bunch of time today to a bizarre bug caused by this and
decided to get rid of it.

Though, instead of *actually* making the claimed change, I decided to
leave in the assertion that I used to catch all of the cases that were
relying on the old behavior.
We can decide later if we want to actually remove the assert.

Supporting the assert required adding a mechanism to reenter a
context.