Make sys:: and cfg:: fit propertly into the object model (#5133)
Objects in `sys` and `cfg` are implemented using custom views.  These
views live in the `edgedbss` module and do not get correctly
incorporated into inheritence views for BaseObject (and even for other
abstract objects in the cfg hierarchy).

Move all of those views into `edgedbstd` with the names they ought to
have. Since we run all the schema DDL before running metaschema
scripts, we need to create dummy views in our DDL, which we do by
doing a `select *` from a dummy table that has the DDL run on it.
Then, when running metaschema, we replace those views with our real
ones.

Doing this means we need to take care to preserve the order of
columns, which is annoying but not hard.

Fixes #4491.

Currently objects in sys:: and cfg::, which