Move introspection from "Guides" to "Schema"
Guides will not be versioned in our documentation. We will separate
versioned sections from non-versions sections in the UI to make it clear
to users which sections are affected by the version selection dropdown
and which are not. Introspection needs to be versioned, so it needs to
be moved out of guides to make it clear once the UI change is
implemented that it is part of the versioned documentation.

This wouldn't have required changing the labels and references to labels
for the introspection section, but I decided to do it anyway to keep the
documentation consistent and to reduce any confusion for users who want
to contribute and see that these labels do not follow our convention.