Update note to reflect bug fix in site watch task
James pointed out that the problem with the build not being updated for
documentation changes was a bug and would be a quick fix. The remaining
gotcha is related to auto-reload which will not trigger on documentation
changes. Updated the note to reflect that remaining issue and to remove
mention of the original issue. A fix will be merged soon for the build
issue.