Infer cardinality of `required multi` pointers as AT_LEAST_ONE (#5180)
Currently we only actually compute the cardinality when accessing
single pointers, and explicitly choose MANY for multi pointers.

I'm not totally sure why it was like this. I think maybe the skeleton
of this logic predated `required multi`? I don't know.

This bug was originally filed last year and was spuriously attributed
to a query builder bug and moved to edgedb/edgedb-js#333, and a query
builder workaround (insert assert_exists calls) was inserted. It was
identified as being a compiler bug, though---but the issue was never
moved back to the edgedb repo and then the workaround got merged
and the issue was closed.