Temporarily(?) exclude sys/cfg views from std::BaseObject (#5181)
* Temporarily(?) exclude sys/cfg views from std::BaseObject

Moving the sys/cfg views into std::BaseObject's inhview (in #5133)
caused catastrophic performance regressions on DDL because of DDL's
reliance on explicit id specification. See #5168.

There is plenty of discussion to be had about what we should do about
this (#5168), but first I want to merge this immediately because the
regression makes our development cycle pretty miserable.

We also move some test config objects under cfg::ConfigObject from
std::Object.

This obviously has to xfail the test that we added in #5133.  Add a
new test that tests some of the other cases that #5133 fixed that we
still support. (Making sure that the abstract parents in sys and cfg
work instead of producing ISEs when used.)