Forbid `on target delete deferred restrict` on required links. (#5189)
We want to forbid on target delete deferred restrict for required links: it
allows intermediate queries to observe the NULL, which breaks the schema's
invariants.

Actually doing that is trivial: just check and throw.

HOWEVER, we were using it a bunch of places in the standard library.

Clean up a bunch of them with the following steps:

Defer if_unused subcommands for object deletions, to try to ensure that the
generated schema types are deleted after the containing objects.
Change @target deletetion handling, by moving it directly to DeletePointer,
cleaning up some cruft between DeleteLink and DeleteProperty in the process.
STILL TODO:

Aliases are a mess. Just punt for aliases and globals, instead marking the
links optional for now.
Fixes #4984.