Support link properties on computed backlinks
Replicating the machinery that non-schema computed backlinks use
(faking inheritance to get the link properties inherited, then undoing
it) turned out to require a huge amount of pain to transplant to a
schema context, so we dropped that whole system and instead check for
link properties on computed backlinks dynamically at compile time.

Fixes #4825.