Make internal trigger names non-obfuscated (#5215)
Currently they are directly produced by get_backend_name by passing in
all the extra stuff as an aspect, which trips us over the length
limit, so the name gets hashed. That length limit doesn't actually
apply to triggers and functions so just produce the name directly.

This will produce a marginal increase in my quality of life when
trying to debug things related to our triggers.