Don't duplicate the computation of single links with link properties
Since single links with link properties are stored both inline in the
object table (for performance, and for compatibility with a potential
parent type that does not have the link property) and in a link table
(where linkprops are always stored), we currently duplicate the
computation when populating them in.

Instead, only do the computation to populate the link table and then
include it into the main table using the rewrites step.

Fixes #4834.