Untangle the inheritance hierachies in the standard library (#5204)
Currently there are a bunch of places where we have inheritance orders
that are not valid C3 orders, because the order bases are specified is
not consistent with the inheritance hierarchies of those bases.

For several, this is just because things are broken in the schema. For
many, it is because the code in reflection that computes how to update
the bases was incorrectly figuring out how to combine the python
orders and the edgeql orders. Typically I fixed those by making
the python mirror the schema better.

Work towards fixing #5118.