Add `ANALYZE` to Edgedb commands. (#5126)
* Add `ANALYZE` to Edgedb commands.

Add a plain `ANALYZE` command without options or arguments.

Issue #4938

* Clean up the grammar.

Make sure that we never parse raw expressions together with non-query
commands. This is to avoid introducing unnecessary reserved keywords for
commands (as expressions can use unreserved keywords as identifiers).

It seems that parsing these together has been a convenience hack rather
than necessity, so it makes sense to roll it back in order to simplify
the grammar.