Enable `pg_dump` functionality.
Provide SQL adapters that allow `pg_dump` to capture the contents of an
EdgeDB database. This functionality is not meant to be used as backup
(it will not be sufficient to restore the original EdgeDB database, use
`edgedb dump` tools for that). It is instead meant to be used as a way
to transfer data to a Postgres database potentially as a compatibility
layer with other SQL tooling.

In the current version the dump completely omits the following:
- pg_stats
- pg_cast
- pg_proc
- pg_operator
- pg_trigger

It also generally modifies some expressions such as defaults because
they may contain references to omitted functions while not affecting the
already stored data.

The user-defined modules get mapped onto namespaces, however the
built-ins and some supporting functionality reside in `edgedbstd` and
`edgedb`.

Co-authored-by: Aljaž Mur Eržen <aljaz.erzen@gmail.com>
Co-authored-by: Michael J. Sullivan <sully@msully.net>