Rename EXPLAIN to ANALYZE, rework syntax
As discussed on offline, arguments will be specified using named tuple
syntax, like
```
analyze (execute := false, buffers := true)
```

Executing the query becomes the default, and can be disabled with
`execute := false`. This is by far the best name for that param that
we've discussed, though using it requires abadoning it as a future
reserved keyword. I feel OK about that.

This turns `ANALYZE` back into a reserved keyword; it was future
reserved in 2.x so this isn't a break.

Fixes #5212.