Require inheritance order to be consistent with the specified base order
It was always, we believe, the intention to implement the C3 ordering
(like Python), and so it was just a bug that this was not enforced in
the past.

Unfortunately this breaks backwards compatability in a pretty obvious
way.

This PR does not do include any backward compatability mechanisms, and
instead will rely on users fixing the issue first and using an
in-development migration squashing tool (edgedb/edgedb-cli#976), the
same as any other bug fix that impacts compatability.

This one might have a bigger impact than most, though, so maybe we do
need to do something?
Or we could just release a beta/RC with it and see if it seems to cause
a lot of grief, and then decide?

Fixes #5118.