Fix errmessage interpolation (#5295)
* Fix errmessage interpolation

Closes #5257

Problem was in errmessage containing json where a JSON object in
curly braces was recognized as an interpolation in Python's `.format`.

Because `errmessage` was not documented to have the whole functionality
that `.format` has, I just replaced `.format` with a regex search and replace.

This will ignore any interpolations that have params (e.g. `{var:02f}`).
It will also not throw InternalServerError when you use interpolations of variables
that were are not provided. This was the reason for this error.

* escaping with double braces and docs