Fix adding certain computed links to a type with an alias
When a type with such a computed pointer that was an alias of another
pointer had an alias created of it, `_elide_derived_ancestors` in the
edgeql compiler was marking the base of the pointer in the view
subclass as being the aliased pointer in the base class, which is
inconsistent with what happened when an ALTER was emitted.

Fix this by having `_elide_derived_ancestors` be more careful about
what it skips over.

Downside of this fix is that it cherry-picking it is dodgy, since it
changes how things are stored after a simple CREATE.