Fix the edgeql+schema patch kind
This had been bitrotted by various changes to our support views.  Be
more limited in which views we recreate, and don't choke when we see
functions.

This never came up because we don't have any patches against the
development branch; it would have been caught by the CI that tests
against old versions once we tried to use it.

Also make it work when AnnotationValue is changed, since the
AnnotationValue object is only exposed in the reflection_schema as
it is reflected as a link in the user exposed schema.

Tested by adding a field to AnnotationValue.