Optimize delta_objects by taking into account name information (#5351)
The idea is simply that if an object named `Foo` exists in both the
old and new schemas, compare those `Foo`s against each other and
nothing else, saving the quadratic cross-product based comparison
for objects that don't appear on both sides.

On the pretty modestly sized `cards` database, this reduces the cost
of the initial `describe current migration as json` for a trivial
property addition from about 3.5s to 0.9s. Dropping the inoperative
`hash_criteria` optimization shaves another 0.05s.

Progress on #5322.