Remove `versionchanged` directive
The information in this block is relevant to v2 and v3 users. As a
result, we don't have an elegant way to show/hide with a version
directive. As a result, we will simplify by just showing it to everyone.
The relevant information about what applies to which version is inline
in the content.