Fix references to single links on `__old__` in triggers
The issue was a subtle condition in path compilation where if a path
was visible, but didn't have a source aspect (like `__old__`), *and*
wasn't in the path_scope (which `__old__` isn't, but also it gets
cleared in DML anyway), then the source rvar we create with
ensure_source_rvar would get lost.

Fixes #5384.