Fix tuples in global variables and do error checks on arrays (#5417)
We need to recode tuples to inject OIDs, like we do with arrays and
the like. Somewhat annoyingly, this ends up as a totally separate code
path from the existing re-encoding of tuples for parameters.

This is because for globals, which are schema declared, we have
postgres composite types, and it would be wasteful and convoluted
to split the tuples up instead of just using them.

And then, once we are injecting the OIDs, we can do missing error
checking that could allow invalid data in. (This was the main thing
that motivated fixing this now instead of deffering.)

Fixes #5356.