Allow `ADMINISTER repair_schema()` outside of test mode (#5421)
Originally "all" ADMINISTER commands (really just ANALYZE) were hidden
behind test mode, but we need to make case-by-case decisions.

I feel pretty strongly that we need to allow `repair_schema()`, though
I would refrain from *documenting* it yet.
One main use case is if we ship a point relase that uses a "repair"
patch, a user takes it and upgrades, and then runs into trouble and
needs to downgrade. Now their schema might be out of sync, and doing
a repair to undo the first repair would solve the problem.

And, importantly, we can't just wait until we've seen users encounter
this problem before doing something, because at that point it would be
too late: they would need the feature present in an *older* version.

But we can skip documenting it until it seems like it is coming up.