Fix making pointer non-computed and giving it an abstract base at the same time
Some obscure details of how reconstruction worked was preventing the alter
from getting joined with the rebase, since the alter had a
`RESET OPTIONALITY` in it. Removed that limitation, and then had to
fix some knock-on effects from that.

Also prohibit having a computed pointer with an abstract base, since
it doesn't work. That check would move this bug up from a migration
time ISE to a real error, though of course we also fixed the bug.

Fixes #5390.