Fix schema ordering bug when INSERTs appear in triggers (#5489)
The problem was that we didn't properly keep track of the path_prefix
when tracing an INSERT, so we didn't catch the dependencies correctly.

While fixing this I also made `alias_context` in the tracer
unconditionally copy the context. Previously we only copied it when
there were aliases, so the path_prefix setting code was modifying an
*outer* context, which is wrong.

Fixes #5481.