Work around mysterious Digital Ocean crashes
When bootstrapping against managed Digital Ocean databases, the
postgres occasionally has a server process crash. This only seems to
happen if we keep our main connection open while initializing the
template DB (!).

So don't do that, I guess.

With this change I've managed 100+ bootstraps in a row without a
failure, so I am pretty confident this is a fix, even though I don't
have a satisfactory explanation of it.