Make staeval more resilient to ops not being supported on a type
staeval supports comparison operators (in order to check constraints
on config values in certain cases) and it supports durations (also for
config reasons). But our duration objects don't support comparison,
so doing a comparison of duration literals crashes in staeval.

This only really happens in the http/notebook protocols, because in
the binary protocol constant extraction breaks any hope of statically
evaluating things.