'Fix' ISE for TypeExprs in function returns (#5540)
We were asserting that we did not use union types in function returns, but all the machinery to create and leverage using those types was already in place. After removing the assert, we were not using the function decls namespace to properly namespace any created types, so do that.

Fixes #5391.