WIP Proof of concept of a pure-rust *concrete* syntax tree parser for edgeql
Basically we just serialize out the state machine, read it in from rust,
and do the shifting and the reducing in a loop.

It produces the CST as json, though obviously there are other options.

It's not /exactly/ pure-rust yet because some meaningful tokenizing
logic occurs in the rust code that translates to python tokens, and
rather than mess with that right now I just run that and then pull the
data back out of the python tokens.
So the lexing+parsing doesn't run any python *code* but it does need
the CPython runtime.