Fix range and bytes output in json-lines/json-pretty mode
The problem was we were always returning jsonb when outputing those types.
This was fine for the regular json mode where it all gets json_agg()d,
but in modes that return multiple json rows this was causing trouble in the
CLI. Properly cast to json.

Unfortunately no test, because I don't think the python bindings support
this mode of operation?

Fixes #5568.