Fix an optional scoping bug with important access policy implications (#5575)
The issue is that in cases like
```
select (
  User,
  (select (
    <str>User.avatar.id ?= '',
    User.name
  )),
);
```
a path for User could get exposed from the ?= computation into the
enclosing computation. This is because `User` doesn't appear under its
enclosing scope in the scope tree (because it is up higher in the tree),
so it doesn't get added to the path id mask.

This is a problem, because if `User.avatar` is empty, then the `?=`
computation will output NULL for everything. This leads `User.name`
to see that NULL, which obliterates the row.

In this case, instead of losing the rows, we compute incorrect values
that result in returning objects when none should be.
```
select (
  Comment.body,
  (select (
    <str>Comment.parent.id ?= '' or
    Comment.id ?= <uuid>{}
  )),
) filter .1;
```

A generalization of this case showed up in a user report where access
policies were exhibiting false positives.

As a hack to suppress certain unsound-for-our-purposes postgres
optimizations, we add `OR .id ?= <uuid>{}` to the filter expressions
we generate to implement policies. This was interacting with the above
bug and was evaluating to *true* in some cases with empty arguments to
optional operators in policies, causing policies to incorrectly pass.

This bug, the `OR .id ?= <uuid>{}` hack, and the prevelance of optional
operations in access policies make this one pretty nasty.

Fix this by masking off any visible and non-optional paths from
optional relations.