Fix several issues related to schema loading in declarative (#5565)
The main change is a new schema where if we can't figure out the type
of an expression in a path, create weak dependencies to all pointers
with the same name.

This doesn't always work but in practice it will do very well.

I have some ideas for how we could incorporate correct typechecking
into schema loading, but in the mean time this probably solves 90% of
the problem at least.

Fixes #4683. Fixes #5163.