Use base schema for migration name lookups (#5579)
We used to use the existing schema for lookups when doing initial ddl type
calculations. This meant that if you referenced a type that existed in the
current schema, but which you intended NOT to exist post-migration, we would get
the bookkeeping wrong. We really want to use the schema for stdlib stuff
anyways. Just use the base.

Fixes #5559.