Swap join waitlist button for link in warning
I don't think the button works great here. It's too close to another
copy of the same button, just up in the navigation. Also, it's awkward
to tell a user in the paragraph above to go down a click the button when
we could just give them a link inline in the paragraph to click.