Add Cloud guide (#5582)
* Add Cloud guide

* Apply suggestions from code review

Co-authored-by: Zack Elan <zackelan@users.noreply.github.com>

* Rename page

* Swap join waitlist button for link in warning

I don't think the button works great here. It's too close to another
copy of the same button, just up in the navigation. Also, it's awkward
to tell a user in the paragraph above to go down a click the button when
we could just give them a link inline in the paragraph to click.

* Clarify interaction modes not mutually exclusive

Co-authored-by: Zack Elan <zackelan@users.noreply.github.com>

* Remove Cloud UI images

Since we will be iterating on this UI quickly

* Add email support option

* Add naming requirements

* Clarify naming length requirement

Co-authored-by: Yury Selivanov <yury@edgedb.com>

* Link to support ticket form

Co-authored-by: Yury Selivanov <yury@edgedb.com>

---------

Co-authored-by: Zack Elan <zackelan@users.noreply.github.com>
Co-authored-by: Yury Selivanov <yury@edgedb.com>