Fix nightly build by removing typing_inspect dependency
Nightlies are failing because:
 * gen_rust_ast.py uses typing_inspect, which is a test dependency
 * The nightly tests run `edb.tool`, which imports all the tools
 * Test dependencies aren't included in the built artifacts
   (tests that use test dependencies are all written to be skipped
   if the dep is missing)

There are a bunch of ways to fix this but the one that I apparently
chose was to make gen_rust_ast instead use the custom simplified
version of typing_inspect that we have for some reason. (Maybe just to
avoid the dependency.) This required fixing it to understand unions
created with `|`.