Fix `pg_namespace` emulation
Current emulation of `pg_namespace` is somewhat broken because it
returns `NULL` for `tableoid` for synthesized schemes, which throws off
catalog map logic in `pg_dump` of PostgreSQL 15.  Emit OID of
`pg_namespace` correctly instead.  While at it, set `xmin`, `xmax`,
`cmin`, `cmax` to non-`NULL` values as well.  `ctid` is harder to
emulate correctly so leave it as `NULL`.