Update desc of `EDGEDB_SERVER_BOOTSTRAP_COMMAND`
Current description says you can use it to create a database and users.
Database creation via this mechanism blocks any other statements from
running since it cannot run in the same block as anything else. This
removes that from the description and offers an alternative for users
who want to create a database and run other statements.