Fix coalesce with type intersection
Closes #5474

The problem was a specific codepath, that was trying to
get_rel_path_output out of a NullRelation for a type intersection
path.

The fix is basically saying "this path does not have storage info",
so the codepath can hit the handler for NullRelation.

This fix was easy write, very hard to find.