Support a primitive form of parameterized types
Types declared in extensions can declare a type scheme that allows
user-defined scalars that extend them to provide the argument value.
Eventually we may want to generalize this to work in arbitrary
situations, but we aren't there yet.

I may have built more generic of infrastructure than necessary: types
declare an array describing the parameters they take and a "type
scheme" that says how to format the modified type names.

But maybe that's just silly and it should just take a number of
parameters, which will need to be integers, and it will format them in
the obvious way?