Fix some obscure optional bugs in the presence of tuple projections
When attaching a path in the scope tree, the detection of whether the
*target* node was optional was insufficient. It only checked the node
itself, but needs to check all the way to the factoring point.
I think that logic had forgotten about BRANCHes that weren't FENCEs.

This came up in a branch where I was experimenting with always
BRANCHing function arguments as a way to validate *sometimes*
branching function arguments to produce code that hits ORDER BY
indexes in complex pgvector cases.