Enable compiling function arguments into subqueries for pgvector opt purposes (#5615)
postgres is finnicky about when it will use indexes for ORDER BY, and
won't do it if the ORDER BY is a subquery. We have machinery to try to
arrange for this to not happen in common cases like `ORDER BY .name`,
but for pgvector, where the clauses might be things like
`euclidean_distance(.embedding, <gpt_vec><str>$0)` or
`euclidean_distance(.embedding, Other.embedding)`,
things get trickier.

We want to generate code that looks approximately like
```
ORDER BY rel.embedding <-> (SELECT ...)
```
rather than what we would typically do, which is something like
```
ORDER BY (
    SELECT rel.embedding <-> q.foo
    FROM (SELECT ... as foo ...) as q
)
```

So we implement a compilation mode for function calls that will try to
put as much of the argument in a direct subquery expression as
possibly. This only works if the argument is actually single, and it
requires compiling the argument under a BRANCH (so that anything
referenced in it is bound in the *argument*, not at the enclosing
scope).

This works, and it passes the test suite.

But I'm wildly uncomfortable with making such a wide-reaching change
to our compilation strategy in 3.0 at such a late date, so I then made
it opt-in on the function level. We'll have the pgvector functions opt
in, and we can consider whether we could and should make this
universal after 3.0