Make Object.property in indexes detached (#5668)
A followup for https://github.com/edgedb/edgedb/pull/5606#issuecomment-1584376671

This is techincally a breaking change, since any index definitions
that are using `Object.property` will now throw errors. I do think it is
quite minor and will probably not affect anyone, as this pattern is not documented
and treated as detached in all other SDL expressions apart from computeds.