Avoid very slow error formatting in non-error cases
When resolve_ptr fails, it does a very slow error enrichment step that
computes levenshtein distances with a ton of different schema
objects. This is fine for actual error cases, but resolve_ptr gets
called in some call sites that swallow the error.

Skip doing the formatting in those cases. This gave a 30% speedup on
some queries.

I did this by checking the `track_ref` variable, since the only cases
where it makes sense to not track references is when we aren't serious
about the lookup, and because adding another param seemed
clunky. Maybe we should just add another, though.