Fix tuple derefencing and top-level casting in exclusive constraints (#5724)
Previously we generated invalid SQL, for two reasons:
 * "Complex" expressions (including the way we emit tuple
   dereference and casting) need to be parenthesized in
   the code we emit to create an EXCLUDE constraint.
 * We were representing the RHS of a tuple dereference
   as a ColumnRef, which got picked up by our system for
   injecting NEW and OLD into trigger expressions,
   which led to us generating bogus expressions like
   (NEW."<some uuid>").NEW."0".

More aggressively parenthesize in constraints, and stop
inappropriately using ColumnRef and use RecordIndirectionOp instead.
(I suspect that when the ast was designed, we used ColumnRef there
as a hack to avoid the bother of adding a pretty marginal
RecordIndirectionOp, which seems to have been added as part of SQL
support, when added all the postgres AST nodes we had skipped).