Fix a casting bug.
Fixes a bug when casting an aliased expression into the same type it
already is.

E.g. `with a := [1] select <array<int64>>a`

The types aren't technically the same, one is Array and the other
ArrayExprAlias, so the cast is not omitted, but then it also cannot be
found because the code assumes that we're casting between different
types.