Back out all the compiler changes, fix it properly in library
No change is actually necessary on the compiler side, but
__range_validate_json needs to be declared as producing OPTIONAL.

The FOR I introduced in the first attempt was just working around not
doing that properly, as I discovered when I tried to go add a comment.

(Honestly, it's arguably a minor performance bug that adding the FOR
helped: we shouldn't emit a NULL check in FOR when the inferred
cardinality is non-empty).