Fix edgeql+schema patches (#5806)
The introduction of edgeql+schema+config patches in #5658 broke
application of the simpler edgeql+schema patches. This is because
certain instdata fields were made versioned, but were only being
updated on edgeql+schema+config patches, which caused them to
not be found after an edgeql+schema patch.

Fix it by unconditionally updating those fields on an edgeql+schema
patch.

This is a forward port of #5804, which I already merged to 3.x.