Avoid recreating ancestor inhviews *twice* on CREATE TYPE (#5814)
Currently we alter ancestor inhviews when we create the initial
inhview for a type. This is pointless: the type hasn't been populated
with pointers like `id` yet, so it won't be included, and the ancestors
will be refreshed in finalize anyway.

This reduces the size of our scripts a bunch, and speeds things up some.