Fix an issue with multiranges.
Not all multiranges are present in the base schema. That causes issues
when they need to be used as array subtypes for parameters as they
cannot be found by ID at all.

This PR makes sure that all base multirange types are part of the
standard schema.

It also fixes JSON serialization of an array of multiranges.