Handle no-default-tenant error in binary protocols
Because the Postgres protocol does TLS handshake differently than TLS
connections, we have to branch into the extracted "handshake" logic if
the `tenant` is not set yet. Because of the lack of tenant, the
"handshake" can only run in an "interruptable" asyncio Task.

EdgeDB binary protocol will always have `tenent` set, unless the SNI is
not matching any tenant. In that case, the "handshake" task will just
serialize and return the proper error from `get_default_tenant()`.