Clarify `if orphan` note on deletion policies
This is an attempt to clear up confusion around `if orphan` in deletion
policies, prompted by a question on Discord:
https://discord.com/channels/841451783728529451/849377751370432573/1133652031635214366

The note was intended to explain that `if orphan` on a link does not
refer to an object being globally orphaned but only with respect to the
link containing the deletion policy. In trying to explain this, the note
ignored the fact that deletion might also be restricted by `on target`
deletion policies of other links linking to the same object, including
the default target deletion policy of `restrict`.

This change caveats the explanation to include the complicating factor
of `on target` deletion policies.