Fix dropping of extensions with nontrivial schemas
The current implementation attempt to clear out the extension innards
directly by issuing `DeleteObject` for every object in the extension
schema.  This isn't entirely correct as it not only can choke `ordering`
with too-granular of an input, but also trips on dropping non-owned
objects unnecessarily.  Avoid this by dropping only the top-level
objects with recursive canonicalization.

Fixes: #5897