Fix `--background`
The `--background` server flag isn't really a recommended way of running
the server (it's best to use a proper service manager like systemd for
that), but sometimes it's useful and we implement it, so fix up the bit
rot, improve logging and add a test.

Note: most of this diff are whitespace changes due to context manager
reshuffling.