Move query builder explanation to top
since that's where it is imported. Trying to handle all pre-requisites
first before we write any code that uses them, so I want to run the
query builder generator and import it there. If we wait until the
end, LSPs will show errors the entire time the user is writing their
code since the generator they imported doesn't exist. This could be
confusing.