Support extension-defined config objects (#5949)
The core user/extension author facing design is that there is a new
config object for extension configs:
```
CREATE ABSTRACT TYPE cfg::ExtensionConfig EXTENDING cfg::ConfigObject {
    CREATE REQUIRED SINGLE LINK cfg -> cfg::AbstractConfig {
        CREATE DELEGATED CONSTRAINT std::exclusive;
    };
};
```
and `AbstractConfig` is updated with a computed back-link:
```
ALTER TYPE cfg::AbstractConfig {
    CREATE MULTI LINK extensions := .<cfg[IS cfg::ExtensionConfig];
};
```

Extension authors can extend ``cfg::ExtensionConfig`` to store
extension specific configuration information. They create a type:
```
CREATE TYPE ext::whatever::WhateverConfig {
    CREATE PROPERTY value -> str { ... }
}
```
The values can then be configured with
```
CONFIGURE CURRENT DATABASE ext::whatever::WhateverConfig::value := ...
```

The config introspection system will generate three `WhateverConfig`
objects, one for each of the `AbstractConfig` subtypes. They can be
fetched directly or via the `extensions` link on
`cfg::AbstractConfig`.

Most of the work in supporting this is in making the config
introspection views understand it properly, and in making sure the
config specs are available where needed.

---

Another change, which probably should have been done separately but
which would be hard to cleanly extract now, is supporting `CONFIGURE
CURRENT DATABASE INSERT`, instead of only supporting it for `INSTANCE`
configs.

This is fairly involved. For `INSTANCE` configs the object-level
operations simply report what changes to make; these changes are then
interpreted by the config system and the entire updated instance
config is written back wholesale as a json blob.
Because we lack the machinery to nicely write back an updated database
config based on changes made in the middle of a script, I instead made
it so that we can do all of the work in SQL, which is kind of a pain because
it is a bunch of jsonb.

I didn't want to try to detect exclusive constraint violations on
those json objects in SQL, so instead we detect the error when
interpreting the returned data from the config operation. This
required adding a new code path to abort

Still TODO (tracked in #5954)
- [ ] Support extension-defined SESSION configs
- [ ] Support redacting config values?
- [ ] Support a second layer of nesting in extension objects
- [ ] See if introspection performance can be improved (cache the get config function in `_edgecon_state`?)
- [ ] Add syntax for "CREATE CONFIG TYPE" to distinguish the types further?
- [ ] Documentation updates?